+++ lib/image_util.rb
@@ -16,4 +16,5 @@ module ImageUtil
   autoload :Terminal, "image_util/terminal"
   autoload :View, "image_util/view"
   autoload :CLI, "image_util/cli"
+  autoload :Benchmarking, "image_util/benchmarking"
 end